+++ CHANGELOG.md
@@ -67,6 +67,7 @@
     - Added a custom soundEffect and custom rpgMusic hook. 
     - Added sound effect for radio button change.
     - Added rpg inspired theme music to play in background of menu. 
+    - Made user input form elements inside the contact/components route.
     - Revamped CHANGELOG.md